+++ CMakeLists.txt
@@ -1,29 +1,12 @@
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.15)
 
-option(BUILD_SAMPLES ON)
+set(ProjectName "mp2-lab3-translator")
 
-set(PROJECT_NAME calculator)
-project(${PROJECT_NAME})
+project(${ProjectName})
 
-include(CTest)
-enable_testing()  # defines BUILD_TESTING
+include_directories(include)
 
-set(MP2_TESTS   "test_${PROJECT_NAME}")
-set(MP2_CUSTOM_PROJECT "${PROJECT_NAME}")
-set(MP2_INCLUDE "${CMAKE_CURRENT_SOURCE_DIR}/include")
-
-add_subdirectory(include)
-
-if(BUILD_TESTING)
-    add_subdirectory(gtest)
-	add_subdirectory(test)
-endif()
-
-
-# REPORT
-message( STATUS "")
-message( STATUS "General configuration for ${PROJECT_NAME}")
-message( STATUS "======================================")
-message( STATUS "")
-message( STATUS "   Configuration: ${CMAKE_BUILD_TYPE}")
-message( STATUS "")
+add_subdirectory(gtest)
+add_subdirectory(run)
+add_subdirectory(src)
+add_subdirectory(test)
