+++ test/test_calculator.cpp
@@ -1,8 +1,9 @@
 #include <gtest.h>
 #include "calculator.h"
-
+/*
 TEST(screen, mew) {
 	screen mew;
 	mew.counting();
 	EXPECT_EQ(1, 1);
 }
+*/
